WIP layouting - Move cytoscape rendering to Drawer
