+++ calm-visualizer/src/components/cytoscape-renderer/cytoscape.css
@@ -5,6 +5,7 @@
     height: 100px;
     font-family: Arial;
     padding: 10px 30px 10px 30px;
+    background-color: white;
 }
 
 .element {
